test: split ModuleDependencyGraph into graph and action
