+++ test/ScanDependencies/Inputs/ModuleDependencyGraph.swift
@@ -144,11 +144,3 @@ struct ModuleDependencyGraph: Codable {
   /// Information about the main module.
   var mainModule: ModuleDependencies { modules[.swift(mainModuleName)]! }
 }
-
-let fileName = CommandLine.arguments[1]
-let data = try! Data(contentsOf: URL(fileURLWithPath: fileName))
-
-let decoder = JSONDecoder()
-let moduleDependencyGraph = try! decoder.decode(
-  ModuleDependencyGraph.self, from: data)
-print(moduleDependencyGraph)
