+++ test/ScanDependencies/Inputs/PrintGraph.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+let fileName = CommandLine.arguments[1]
+let data = try! Data(contentsOf: URL(fileURLWithPath: fileName))
+
+let decoder = JSONDecoder()
+let moduleDependencyGraph = try! decoder.decode(
+  ModuleDependencyGraph.self, from: data)
+print(moduleDependencyGraph)
